Update
git-svn-id: https://svn.l4re.org/repos/oc/l4re/trunk@73 d050ee49-bd90-4346-b210-929a50b99cfc

+++ l4/pkg/spafs/server/src/main.cc
@@ -180,7 +180,7 @@ Fprov_server::get_file(char const *filename, unsigned len, L4::Ipc::Snd_fpage &r
       return -L4_ENODEV;
     }
 
-  if (file_ds->size() != (long)dir_entry->size)
+  if (file_ds->size() != dir_entry->size)
     printf("Warning: file-dataspace has different size from file\n");
 
   if (file_ds->copy_in(0, _ds, dir_entry->addr, dir_entry->size))
